migrate inventory quantities from int to int64

+++ services/inventory/internal/domain/model.go
@@ -7,16 +7,16 @@ import (
 type Inventory struct {
 	LocationID string    `json:"location_id"`
 	SKU        string    `json:"sku"`
-	OnHand     int       `json:"on_hand"`
-	Reserved   int       `json:"reserved"`
-	Incoming   int       `json:"incoming"`
+	OnHand     int64     `json:"on_hand"`
+	Reserved   int64     `json:"reserved"`
+	Incoming   int64     `json:"incoming"`
 	UpdatedAt  time.Time `json:"updated_at"`
 }
 
 type UpdatedInventory struct {
 	LocationID *string `json:"location_id"`
 	SKU        *string `json:"sku"`
-	OnHand     *int    `json:"on_hand"`
-	Reserved   *int    `json:"reserved"`
-	Incoming   *int    `json:"incoming"`
+	OnHand     *int64  `json:"on_hand"`
+	Reserved   *int64  `json:"reserved"`
+	Incoming   *int64  `json:"incoming"`
 }
